set cache control header on gem serving
if we're public use a year, if private cache for 1 hour.

we may want to tweak these values based on fastly.

+++ test/controllers/namespaces/gems_controller_test.rb
@@ -20,7 +20,7 @@ class Namespaces::GemsControllerTest < ActionDispatch::IntegrationTest
     upload = Peak::Gem::Upload.read(StringIO.new(response.body))
     assert upload.spec
   ensure
-    upload.unlink
+    upload.unlink if upload # Guard against never reaching the assignment line
   end
 
   test "push" do
